requirements.txt
# Основные библиотеки
python-dotenv>=1.0.0
openai>=1.3.0
chromadb>=0.4.15

# Для эмбеддингов
sentence-transformers>=2.2.2

# Для интерактивного ввода
prompt_toolkit>=3.0.36

# Для работы с API
requests>=2.31.0

# Для логирования и обработки данных
tqdm>=4.65.0
numpy>=1.24.0

# Для работы с ОС и путями
pathlib>=1.0.1 